refactor: split largest screens, fix latent player + cache issues
Phase 1 — latent fixes & API modernization:
- NativeVideoPresenter: invalidate KVO observation before reassigning on
  episode nav (was leaking observers on every prev/next).
- AppNavigation: configure Nuke ImageCache with 256 MB cost limit so 4K
  backdrops stop evicting mid-render on tvOS.
- TrackPickerSheet: NavigationView -> NavigationStack.
- FocusScaleModifier: replace EnvironmentKey boilerplate with @Entry macro.
- SettingsRowHelpers: @MainActor on serverStatusBadge / iOSSettingsRow /
  iOSRowIcon to comply with the documented Swift 6 rule.

Phase 2 — MediaDetailScreen split (1122 -> 777 LOC):
Extract 7 sibling files (EpisodeCard, EpisodeRow, EpisodeMetadataLine,
EpisodeOverviewSheet, CastSection, SimilarSection, ButtonStyles). Each
View struct is Equatable (mirroring the existing PlayActionButtonsSection
pattern) so SwiftUI short-circuits re-renders on unrelated parent
mutations. The tvOS episode list now uses LazyVStack. When `==` reads
non-Sendable BaseItemDto fields, body is wrapped in
MainActor.assumeIsolated — documented as escape hatch #4 in CLAUDE.md.

Phase 3 — Settings/ subfolder + tvOS extractions (1000 -> 631 LOC):
git-mv 5 settings files into Shared/Screens/Settings/ and split out
SettingsTVAccentPicker / SettingsTVLanguagePicker / SettingsTVProfileSection
/ SettingsTVActionRow as extension files. Bloom magic numbers promoted
to CinemaBloom tokens.

Phase 4 — RemoteCommandController extracted into VideoPlayer/, mirroring
the existing sub-controller pattern (PlaybackReporter / SkipSegment /
SleepTimer / Chapter / EndOfSeriesOverlay). Presenter now calls
remoteCommands.attach/detach instead of inline setup.

Phase 5 — networking + screen polish:
- JellyfinAPIClient+Playback: extract resolvePlayableEpisode helper
  (Series/Season -> Episode resolution path).
- PrivacySecurityScreen: lift ConnectedDevicesList to its own file
  (501 -> 320 LOC).

Behavior unchanged. Both iOS + tvOS targets build clean (set -o pipefail
+ ** BUILD SUCCEEDED ** on iPhone 17 Pro / Apple TV 4K 3rd gen).
