+++ Shared/Screens/Settings/SettingsScreen+tvOS.swift
@@ -44,8 +44,8 @@ extension SettingsScreen {
             // Centered accent bloom — persists across all settings pages
             Circle()
                 .fill(themeManager.accent.opacity(0.3))
-                .frame(width: 1200, height: 1200)
-                .blur(radius: 280)
+                .frame(width: CinemaBloom.settingsSize, height: CinemaBloom.settingsSize)
+                .blur(radius: CinemaBloom.settingsBlur)
         }
         .task { await loadServerUsers() }
         .alert(loc.localized("action.logOut"), isPresented: $showLogOutAlert) {
@@ -428,291 +428,6 @@ extension SettingsScreen {
         }
     }
 
-    var tvLanguagePicker: some View {
-        let isFocused = focusedItem == .language("row")
-
-        return Button {
-            // Toggle between languages on press
-            loc.languageCode = loc.languageCode == "fr" ? "en" : "fr"
-        } label: {
-            HStack(spacing: CinemaSpacing.spacing3) {
-                Image(systemName: "globe")
-                    .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                    .foregroundStyle(themeManager.accent)
-                    .frame(width: 24)
-
-                Text(loc.localized("settings.language"))
-                    .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                    .foregroundStyle(CinemaColor.onSurface)
-
-                Spacer()
-
-                HStack(spacing: CinemaSpacing.spacing2) {
-                    tvLanguageChip("fr", label: loc.localized("settings.language.french"))
-                    tvLanguageChip("en", label: loc.localized("settings.language.english"))
-                }
-            }
-            .padding(.horizontal, CinemaSpacing.spacing4)
-            .frame(maxWidth: .infinity, minHeight: 80)
-            .tvSettingsFocusable(isFocused: isFocused, accent: themeManager.accent, animated: motionEffects, colorScheme: themeManager.darkModeEnabled ? .dark : .light)
-        }
-        .buttonStyle(.plain)
-        .focusEffectDisabled()
-        .hoverEffectDisabled()
-        .focused($focusedItem, equals: .language("row"))
-        .onMoveCommand { direction in
-            guard isFocused else { return }
-            switch direction {
-            case .left, .right:
-                loc.languageCode = loc.languageCode == "fr" ? "en" : "fr"
-            default:
-                break
-            }
-        }
-    }
-
-    func tvLanguageChip(_ code: String, label: String) -> some View {
-        let isSelected = loc.languageCode == code
-
-        return Text(label)
-            .font(.system(size: CinemaScale.pt(20), weight: isSelected ? .bold : .medium))
-            .foregroundStyle(isSelected ? themeManager.onAccent : CinemaColor.onSurfaceVariant)
-            .padding(.horizontal, CinemaSpacing.spacing4)
-            .padding(.vertical, CinemaSpacing.spacing2)
-            .background(
-                RoundedRectangle(cornerRadius: CinemaRadius.medium)
-                    .fill(isSelected ? themeManager.accent : CinemaColor.surfaceContainerHigh)
-            )
-    }
-
-    var tvAccentColorPicker: some View {
-        let isFocused = focusedItem == .accentColor("row")
-        let allOptions = AccentOption.visibleCases(rainbowUnlocked: rainbowUnlocked)
-
-        return Button {
-            // Cycle to next accent color on press
-            if let currentIndex = allOptions.firstIndex(where: { $0.rawValue == themeManager.accentColorKey }) {
-                let nextIndex = (allOptions.distance(from: allOptions.startIndex, to: currentIndex) + 1) % allOptions.count
-                themeManager.accentColorKey = allOptions[allOptions.index(allOptions.startIndex, offsetBy: nextIndex)].rawValue
-            }
-        } label: {
-            HStack(spacing: CinemaSpacing.spacing3) {
-                Image(systemName: "paintpalette.fill")
-                    .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                    .foregroundStyle(themeManager.accent)
-                    .frame(width: 24)
-
-                Text(loc.localized("settings.accentColor"))
-                    .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                    .foregroundStyle(CinemaColor.onSurface)
-
-                Spacer()
-
-                HStack(spacing: CinemaSpacing.spacing3) {
-                    ForEach(allOptions) { option in
-                        tvAccentDotDisplay(option)
-                    }
-                }
-            }
-            .padding(.horizontal, CinemaSpacing.spacing4)
-            .frame(maxWidth: .infinity, minHeight: 80)
-            .tvSettingsFocusable(isFocused: isFocused, accent: themeManager.accent, animated: motionEffects, colorScheme: themeManager.darkModeEnabled ? .dark : .light)
-        }
-        .buttonStyle(.plain)
-        .focusEffectDisabled()
-        .hoverEffectDisabled()
-        .focused($focusedItem, equals: .accentColor("row"))
-        .onMoveCommand { direction in
-            guard isFocused else { return }
-            let allOpts = AccentOption.visibleCases(rainbowUnlocked: rainbowUnlocked)
-            if let currentIndex = allOpts.firstIndex(where: { $0.rawValue == themeManager.accentColorKey }) {
-                let idx = allOpts.distance(from: allOpts.startIndex, to: currentIndex)
-                switch direction {
-                case .left:
-                    if idx > 0 {
-                        themeManager.accentColorKey = allOpts[allOpts.index(allOpts.startIndex, offsetBy: idx - 1)].rawValue
-                    }
-                case .right:
-                    if idx < allOpts.count - 1 {
-                        themeManager.accentColorKey = allOpts[allOpts.index(allOpts.startIndex, offsetBy: idx + 1)].rawValue
-                    }
-                default:
-                    break
-                }
-            }
-        }
-    }
-
-    func tvAccentDotDisplay(_ option: AccentOption) -> some View {
-        let isSelected = option.rawValue == themeManager.accentColorKey
-
-        return ZStack {
-            if option == .rainbow {
-                RainbowAccentSwatch(diameter: 36)
-            } else {
-                Circle()
-                    .fill(option.color)
-                    .frame(width: 36, height: 36)
-            }
-
-            if isSelected {
-                Circle()
-                    .strokeBorder(.white.opacity(0.9), lineWidth: 2.5)
-                    .frame(width: 36, height: 36)
-
-                Image(systemName: "checkmark")
-                    .font(.system(size: CinemaScale.pt(14), weight: .bold))
-                    .foregroundStyle(.white)
-            }
-        }
-        .frame(width: 48, height: 48)
-    }
-
-    // MARK: - Profile Management
-
-    var currentUserId: String {
-        appState.currentUserId ?? ""
-    }
-
-    var displayUsers: [UserDto] {
-        let users: [UserDto]
-        if serverUsers.isEmpty {
-            if let session = appState.keychain.getUserSession() {
-                users = [UserDto(id: session.userID, name: session.username)]
-            } else {
-                users = []
-            }
-        } else {
-            users = serverUsers
-        }
-        return users.sorted { a, _ in a.id == currentUserId }
-    }
-
-    var tvProfileSection: some View {
-        VStack(alignment: .leading, spacing: CinemaSpacing.spacing3) {
-            tvSectionLabel(loc.localized("settings.profiles"))
-
-            LazyVGrid(
-                columns: [
-                    GridItem(.flexible(), spacing: CinemaSpacing.spacing3),
-                    GridItem(.flexible(), spacing: CinemaSpacing.spacing3),
-                    GridItem(.flexible(), spacing: CinemaSpacing.spacing3)
-                ],
-                spacing: CinemaSpacing.spacing3
-            ) {
-                ForEach(displayUsers, id: \.id) { user in
-                    tvProfileBlock(user: user)
-                }
-
-                tvSwitchAccountBlock
-            }
-        }
-    }
-
-    func tvProfileBlock(user: UserDto) -> some View {
-        let userId = user.id ?? ""
-        let isCurrentUser = userId == currentUserId
-        let hasImage = user.primaryImageTag != nil
-        let isFocused = focusedItem == .profile(userId)
-
-        return Button {
-            if !isCurrentUser { showUserSwitch = true }
-        } label: {
-            VStack(spacing: CinemaSpacing.spacing2) {
-                Group {
-                    if hasImage, appState.serverURL != nil {
-                        let imageURL = appState.imageBuilder
-                            .userImageURL(userId: userId, tag: user.primaryImageTag, maxWidth: 96)
-                        // Initial serves as placeholder (while loading) and
-                        // fallback (on 404); CinemaLazyImage draws above it
-                        // once the image resolves (fallbackBackground is clear
-                        // so the initial shows through on transient states).
-                        tvUserInitial(name: user.name ?? "?", size: 36)
-                            .overlay {
-                                CinemaLazyImage(
-                                    url: imageURL,
-                                    fallbackIcon: nil,
-                                    fallbackBackground: .clear
-                                )
-                            }
-                            .frame(width: 36, height: 36)
-                            .clipShape(Circle())
-                    } else {
-                        tvUserInitial(name: user.name ?? "?", size: 36)
-                    }
-                }
-                .opacity(isCurrentUser ? 1.0 : 0.55)
-
-                Text(user.name ?? "User")
-                    .font(.system(size: CinemaScale.pt(17), weight: .semibold))
-                    .foregroundStyle(isFocused ? CinemaColor.onSurface : CinemaColor.onSurfaceVariant)
-                    .lineLimit(1)
-
-                if isCurrentUser {
-                    Text(loc.localized("settings.active"))
-                        .font(.system(size: CinemaScale.pt(13), weight: .bold))
-                        .foregroundStyle(CinemaColor.success)
-                } else {
-                    Text(user.policy?.isAdministrator == true ? loc.localized("settings.admin") : loc.localized("settings.user"))
-                        .font(.system(size: CinemaScale.pt(13), weight: .medium))
-                        .foregroundStyle(isFocused ? CinemaColor.onSurface : CinemaColor.onSurfaceVariant)
-                }
-            }
-            .frame(maxWidth: .infinity, minHeight: 80)
-            .padding(.vertical, CinemaSpacing.spacing3)
-            .tvSettingsFocusable(isFocused: isFocused, accent: themeManager.accent, colorScheme: themeManager.darkModeEnabled ? .dark : .light)
-        }
-        .buttonStyle(.plain)
-        .focusEffectDisabled()
-        .hoverEffectDisabled()
-        .focused($focusedItem, equals: .profile(userId))
-    }
-
-    var tvSwitchAccountBlock: some View {
-        let isFocused = focusedItem == .switchAccount
-
-        return Button {
-            showUserSwitch = true
-        } label: {
-            VStack(spacing: CinemaSpacing.spacing2) {
-                Image(systemName: "arrow.left.arrow.right")
-                    .font(.system(size: CinemaScale.pt(20), weight: .semibold))
-                    .foregroundStyle(isFocused ? CinemaColor.onSurface : CinemaColor.onSurfaceVariant)
-                    .frame(width: 36, height: 36)
-
-                Text(loc.localized("settings.switchAccount"))
-                    .font(.system(size: CinemaScale.pt(17), weight: .semibold))
-                    .foregroundStyle(isFocused ? CinemaColor.onSurface : CinemaColor.onSurfaceVariant)
-                    .multilineTextAlignment(.center)
-            }
-            .frame(maxWidth: .infinity, minHeight: 80)
-            .padding(.vertical, CinemaSpacing.spacing3)
-            .tvSettingsFocusable(isFocused: isFocused, accent: themeManager.accent, colorScheme: themeManager.darkModeEnabled ? .dark : .light)
-        }
-        .buttonStyle(.plain)
-        .focusEffectDisabled()
-        .hoverEffectDisabled()
-        .focused($focusedItem, equals: .switchAccount)
-    }
-
-    func tvUserInitial(name: String, size: CGFloat) -> some View {
-        ZStack {
-            Circle()
-                .fill(
-                    LinearGradient(
-                        colors: [themeManager.accentContainer, themeManager.accent.opacity(0.5)],
-                        startPoint: .topLeading,
-                        endPoint: .bottomTrailing
-                    )
-                )
-                .frame(width: size, height: size)
-
-            Text(String(name.prefix(1)).uppercased())
-                .font(.system(size: size * 0.42, weight: .bold))
-                .foregroundStyle(.white)
-        }
-    }
-
     // MARK: - Server Connection
 
     var tvRefreshConnectionButton: some View {
@@ -881,90 +596,6 @@ extension SettingsScreen {
         .focusEffectDisabled()
         .hoverEffectDisabled()
         .focused($focusedItem, equals: .toggle(key))
-    }
-
-    // MARK: - Action Row (shared tvOS helper)
-    //
-    // Single source of truth for "tappable row with icon + title + optional
-    // subtitle + optional chevron" on tvOS. Replaces three near-duplicate
-    // bespoke buttons (Refresh Catalogue / Refresh Connection / Licenses).
-    //
-    // Two overloads let callers either reuse the generic `.toggle(id)` focus
-    // lane (used for most settings rows) or supply a dedicated `SettingsFocus`
-    // case where one already exists (e.g. `.refreshConnection`).
-
-    @ViewBuilder
-    func tvActionRow(
-        id: String,
-        icon: String,
-        label: String,
-        subtitle: String? = nil,
-        showsChevron: Bool = false,
-        tint: Color? = nil,
-        action: @escaping () -> Void
-    ) -> some View {
-        tvActionRow(
-            focus: .toggle(id),
-            icon: icon,
-            label: label,
-            subtitle: subtitle,
-            showsChevron: showsChevron,
-            tint: tint,
-            action: action
-        )
-    }
-
-    @ViewBuilder
-    func tvActionRow(
-        focus: SettingsFocus,
-        icon: String,
-        label: String,
-        subtitle: String? = nil,
-        showsChevron: Bool = false,
-        tint: Color? = nil,
-        action: @escaping () -> Void
-    ) -> some View {
-        let isFocused = focusedItem == focus
-        let iconColor = tint ?? themeManager.accent
-        let labelColor = tint ?? CinemaColor.onSurface
-        Button(action: action) {
-            HStack(spacing: CinemaSpacing.spacing3) {
-                Image(systemName: icon)
-                    .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                    .foregroundStyle(iconColor)
-                    .frame(width: 24)
-
-                if let subtitle {
-                    VStack(alignment: .leading, spacing: 4) {
-                        Text(label)
-                            .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                            .foregroundStyle(labelColor)
-                        Text(subtitle)
-                            .font(.system(size: CinemaScale.pt(16), weight: .regular))
-                            .foregroundStyle(CinemaColor.onSurfaceVariant)
-                    }
-                } else {
-                    Text(label)
-                        .font(.system(size: CinemaScale.pt(20), weight: .medium))
-                        .foregroundStyle(labelColor)
-                }
-
-                Spacer()
-
-                if showsChevron {
-                    Image(systemName: "chevron.right")
-                        .font(.system(size: CinemaScale.pt(14), weight: .medium))
-                        .foregroundStyle(CinemaColor.onSurfaceVariant)
-                }
-            }
-            .padding(.horizontal, CinemaSpacing.spacing4)
-            .frame(maxWidth: .infinity, minHeight: 80)
-            .tvSettingsFocusable(isFocused: isFocused, accent: themeManager.accent, colorScheme: themeManager.darkModeEnabled ? .dark : .light)
-        }
-        .buttonStyle(.plain)
-        .focusEffectDisabled()
-        .hoverEffectDisabled()
-        .focused($focusedItem, equals: focus)
     }
 
     // MARK: - tvOS Helpers
